+++ changelog.md
@@ -1,3 +1,11 @@
+# Version 0.3
+- Connections display in main window
+- Create a connection
+  - Use tlclient_template.conf as a template
+- Launch connection from main window
+  - tlclient saved config will be preserved
+- Save connections to connections.json
+
 # Version 0.2
 - Added ThinLinc client detection for Linux